LibGfx: Move ScaledFont and new base class VectorFont out of TTF
