+++ Userland/Libraries/LibGfx/Font/TrueType/Font.cpp
@@ -8,8 +8,6 @@
 
 #include <AK/Checked.h>
 #include <AK/Try.h>
-#include <AK/Utf32View.h>
-#include <AK/Utf8View.h>
 #include <LibCore/MappedFile.h>
 #include <LibGfx/Font/TrueType/Cmap.h>
 #include <LibGfx/Font/TrueType/Font.h>
@@ -526,13 +524,13 @@ ErrorOr<NonnullRefPtr<Font>> Font::try_load_from_offset(ReadonlyBytes buffer, u3
     return adopt_ref(*new Font(move(buffer), move(head), move(name), move(hhea), move(maxp), move(hmtx), move(cmap), move(loca), move(glyf), move(os2), move(kern)));
 }
 
-ScaledFontMetrics Font::metrics([[maybe_unused]] float x_scale, float y_scale) const
+Gfx::ScaledFontMetrics Font::metrics([[maybe_unused]] float x_scale, float y_scale) const
 {
     auto ascender = m_hhea.ascender() * y_scale;
     auto descender = m_hhea.descender() * y_scale;
     auto line_gap = m_hhea.line_gap() * y_scale;
 
-    return ScaledFontMetrics {
+    return Gfx::ScaledFontMetrics {
         .ascender = ascender,
         .descender = descender,
         .line_gap = line_gap,
@@ -540,7 +538,7 @@ ScaledFontMetrics Font::metrics([[maybe_unused]] float x_scale, float y_scale) c
 }
 
 // FIXME: "loca" and "glyf" are not available for CFF fonts.
-ScaledGlyphMetrics Font::glyph_metrics(u32 glyph_id, float x_scale, float y_scale) const
+Gfx::ScaledGlyphMetrics Font::glyph_metrics(u32 glyph_id, float x_scale, float y_scale) const
 {
     if (glyph_id >= glyph_count()) {
         glyph_id = 0;
@@ -550,7 +548,7 @@ ScaledGlyphMetrics Font::glyph_metrics(u32 glyph_id, float x_scale, float y_scal
     auto glyph = m_glyf.glyph(glyph_offset);
     int ascender = glyph.ascender();
     int descender = glyph.descender();
-    return ScaledGlyphMetrics {
+    return Gfx::ScaledGlyphMetrics {
         .ascender = (int)roundf(ascender * y_scale),
         .descender = (int)roundf(descender * y_scale),
         .advance_width = (int)roundf(horizontal_metrics.advance_width * x_scale),
@@ -642,100 +640,6 @@ bool Font::is_fixed_width() const
     // FIXME: Read this information from the font file itself.
     // FIXME: Although, it appears some application do similar hacks
     return glyph_metrics(glyph_id_for_code_point('.'), 1, 1).advance_width == glyph_metrics(glyph_id_for_code_point('X'), 1, 1).advance_width;
-}
-
-int ScaledFont::width(StringView view) const { return unicode_view_width(Utf8View(view)); }
-int ScaledFont::width(Utf8View const& view) const { return unicode_view_width(view); }
-int ScaledFont::width(Utf32View const& view) const { return unicode_view_width(view); }
-
-template<typename T>
-ALWAYS_INLINE int ScaledFont::unicode_view_width(T const& view) const
-{
-    if (view.is_empty())
-        return 0;
-    int width = 0;
-    int longest_width = 0;
-    u32 last_code_point = 0;
-    for (auto code_point : view) {
-        if (code_point == '\n' || code_point == '\r') {
-            longest_width = max(width, longest_width);
-            width = 0;
-            last_code_point = code_point;
-            continue;
-        }
-        u32 glyph_id = glyph_id_for_code_point(code_point);
-        auto kerning = glyphs_horizontal_kerning(last_code_point, code_point);
-        width += kerning + glyph_metrics(glyph_id).advance_width;
-        last_code_point = code_point;
-    }
-    longest_width = max(width, longest_width);
-    return longest_width;
-}
-
-RefPtr<Gfx::Bitmap> ScaledFont::rasterize_glyph(u32 glyph_id) const
-{
-    auto glyph_iterator = m_cached_glyph_bitmaps.find(glyph_id);
-    if (glyph_iterator != m_cached_glyph_bitmaps.end())
-        return glyph_iterator->value;
-
-    auto glyph_bitmap = m_font->rasterize_glyph(glyph_id, m_x_scale, m_y_scale);
-    m_cached_glyph_bitmaps.set(glyph_id, glyph_bitmap);
-    return glyph_bitmap;
-}
-
-Gfx::Glyph ScaledFont::glyph(u32 code_point) const
-{
-    auto id = glyph_id_for_code_point(code_point);
-    auto bitmap = rasterize_glyph(id);
-    auto metrics = glyph_metrics(id);
-    return Gfx::Glyph(bitmap, metrics.left_side_bearing, metrics.advance_width, metrics.ascender);
-}
-
-u8 ScaledFont::glyph_width(u32 code_point) const
-{
-    auto id = glyph_id_for_code_point(code_point);
-    auto metrics = glyph_metrics(id);
-    return metrics.advance_width;
-}
-
-int ScaledFont::glyph_or_emoji_width(u32 code_point) const
-{
-    auto id = glyph_id_for_code_point(code_point);
-    auto metrics = glyph_metrics(id);
-    return metrics.advance_width;
-}
-
-float ScaledFont::glyphs_horizontal_kerning(u32 left_code_point, u32 right_code_point) const
-{
-    if (left_code_point == 0 || right_code_point == 0)
-        return 0.f;
-
-    auto left_glyph_id = glyph_id_for_code_point(left_code_point);
-    auto right_glyph_id = glyph_id_for_code_point(right_code_point);
-    if (left_glyph_id == 0 || right_glyph_id == 0)
-        return 0.f;
-
-    return m_font->glyphs_horizontal_kerning(left_glyph_id, right_glyph_id, m_x_scale);
-}
-
-u8 ScaledFont::glyph_fixed_width() const
-{
-    return glyph_metrics(glyph_id_for_code_point(' ')).advance_width;
-}
-
-Gfx::FontPixelMetrics ScaledFont::pixel_metrics() const
-{
-    auto metrics = m_font->metrics(m_x_scale, m_y_scale);
-
-    return Gfx::FontPixelMetrics {
-        .size = (float)pixel_size(),
-        .x_height = (float)x_height(),
-        .advance_of_ascii_zero = (float)glyph_width('0'),
-        .glyph_spacing = (float)glyph_spacing(),
-        .ascent = metrics.ascender,
-        .descent = -metrics.descender,
-        .line_gap = metrics.line_gap,
-    };
 }
 
 u16 OS2::weight_class() const
